Generate Kubebuilder created types
Remove Extra Kubebuilder Bits

For now this repo will serve to only house the
AdminNetworkPolicy API, while implementation of the
API will happen within variou CNIs. Therefore
this repo will not ship any binaries related to the
ANP API, allowing for all of the contollrer managment
bits to be removed.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>