Leverage kubebuilder annotations generator
This PR updates the controller-tools version to 0.11.1.
That will give us the support for annotating CRDs with
api-approved annotation using kubebuilder. See
https://github.com/kubernetes-sigs/controller-tools/issues/751.

This is a much cleaner way of generating the right annotations.
Hence we don't need to do the extra patches and can remove
some redundant files.

This makes lifes easier for CNI plugin implementations as
they try to install the CRDs since vendoring the repo
will not bring in the `config` folder and all plugins
will need to generate their own set of CRDs. Hence its
better to annotate this using kubebuilder to avoid
expecting plugins to patch the annotation manually

Solves 2 & 3 points in
https://github.com/kubernetes-sigs/network-policy-api/issues/76

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>