Add validation for allowing only 1 BANP
The KEP says that at a given time, we can have only
1 BANP in the cluster. No use cases for more than one
was brough forth so far. So in the v1alpha1 of the API
let's stick to this and we can later remove this
restriction if more than one BANP is required, but then
that will entail confusions since BANP doesn't have a
priority set.

Since there is no straight forward way to limit
Items in the BaselineAdminNetworkPolicyList, we will use
the new xvalidation:
https://github.com/kubernetes/enhancements/blob/master/keps/sig-api-machinery/2876-crd-validation-expression-language/README.md#proposal
approach here. See how singleton is expressed.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>