Remove standard from feature-set names
We were calling our feature sets as
StandardCoreFeatures and StandardExtendedFeatures.
This was in anticipation of Standard and experimental
channels which is not something we have strongly started
to use.

So having standard and extended as way forward is easier
than having standard.core, standard.extended and
experimental.core, experimental.extended sets.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>