fixup the api-approved annotation
The previous pr in our api-approved.kubernetes.io annotation
did not include a review from the members of the `sig-network-api-approvers`
group.  This PR is the original API implementation PR that was reviewed/approved
by Tim Hockin, the core of the API hasn't really chaged since then. For
future releases with major API changes we will ensure that the API is
properly reviewed.

Signed-off-by: astoycos <astoycos@redhat.com>