Adds unit test to test CRDs
Tests the current CRDs against example valid and invalid CRDs. There
are two ways to use this:

Test cases
----------

Each test case takes its name from the name of the resource being
instantiated.

 - pkg/crdtest/testdata/{valid, invalid}

Running
-------

Run as a standard Go-test:

  $ go test ./pkg/crdtest

Run in -watch mode for development:

  $ go test -c -o crdtest ./pkg/crdtest # or "make crdtest"
  $ ./crdtest -crdDir config/crd/standard -watch pkg/crdtest/testdata/valid

will apply all yamls in pkg/crdtest/valid folder and continue watching
the changes to the *.yaml files in the folder. When a file is
changed, it will load it into an API server and print the result.